Temp solution to drive progress bar
Will run robocopy with /l flag that generates report ( no copy ) , then will read / clean that text file and add the files into a list. Then when actual copying is happening, run while loop that checks if item in list is in destination. If so, pops it out of the list and increase the progress bar