Working on troubleshooting progress bar
This feature is becoming a real headache, I took a step back to looked at all the variables and defined a clear path to record the progress, and it's still not working as it should. I'm considering disabling this feature for now as this time taking up too much time.